backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
pytesseract
pillow
openai-whisper
torch
numpy
requests
